Updated java interface.
Summary:
- fixed some visibility issues
- added magic constants that we use to determine first seqno when subscribing to the topic
- changed C++ <-> Java mapping to uint16_t <-> int, since coding it in short was just bad

Test Plan: `mvn test`

Reviewers: pja, dhruba, ranji42

Reviewed By: ranji42

Differential Revision: https://reviews.facebook.net/D31659

+++ src/main/java/org/rocketspeed/MessageReceived.java
@@ -2,20 +2,19 @@ package org.rocketspeed;
 
 public class MessageReceived {
 
-  private final short mNamespaceId;
+  private final int mNamespaceId;
   private final String mTopicName;
   private final long mSequenceNumber;
   private final byte[] mContents;
 
-  public MessageReceived(short namespaceId, String topicName, long sequenceNumber,
-                         byte[] contents) {
+  public MessageReceived(int namespaceId, String topicName, long sequenceNumber, byte[] contents) {
     this.mNamespaceId = namespaceId;
     this.mTopicName = topicName;
     this.mSequenceNumber = sequenceNumber;
     this.mContents = contents;
   }
 
-  public short getNamespaceId() {
+  public int getNamespaceId() {
     return mNamespaceId;
   }
