+++ src/main/java/org/rocketspeed/PublishCallback.java
@@ -2,6 +2,6 @@ package org.rocketspeed;
 
 public interface PublishCallback {
 
-  void call(Status status, short namespaceId, String topicName, MsgId messageId,
-            long sequenceNumber, byte[] contents);
+  void call(Status status, int namespaceId, String topicName, MsgId messageId, long sequenceNumber,
+            byte[] contents);
 }
